Update dependencies and enhance documentation for text splitters and loaders
- Updated `text-splitter` dependency to version 0.29 and added new features for code parsing.
- Introduced new optional dependencies in `Cargo.toml` for YAML, TOML, XML, Excel, and base64 support.
- Enhanced README files with detailed descriptions of text splitters and document loaders, including new examples for various file types and retrieval methods.
- Improved error handling in document loaders to accommodate new formats and added relevant error types.
- Refactored text splitter module to include additional splitters for HTML and JSON, and updated the options for chunk sizing.

+++ src/document_loaders/source_code_loader/language_parsers.rs
@@ -156,7 +156,7 @@ impl LanguageParser {
 
         let count = tree.root_node().child_count();
         for i in 0..count {
-            let node = tree.root_node().child(i).unwrap();
+            let node = tree.root_node().child(i as u32).unwrap();
             let source_code = node.utf8_text(code.as_bytes()).unwrap().to_string();
             let lang_meta = (
                 "language".to_string(),
